feat(shared): change int to long for id

+++ SmartIndustries.Smartlock.Platform/Shared/Domain/Repositories/IBaseRepository.cs
@@ -3,7 +3,7 @@ namespace SmartIndustries.Smartlock.Platform.Shared.Domain.Repositories;
 public interface IBaseRepository<TEntity>
 {
     Task AddAsync(TEntity entity, CancellationToken cancellationToken = default);
-    Task<TEntity?> FindByIdAsync(int id, CancellationToken cancellationToken = default);
+    Task<TEntity?> FindByIdAsync(long id, CancellationToken cancellationToken = default);
     void Update(TEntity entity);
     void Remove(TEntity entity);
     Task<IEnumerable<TEntity>> ListAsync(CancellationToken cancellationToken = default);
